+++ things/postgres/init.go
@@ -34,7 +34,7 @@ func migrateDB(db *sql.DB) error {
 				Id: "things_1",
 				Up: []string{
 					`CREATE TABLE things (
-						id      SERIAL,
+						id      BIGSERIAL,
 						owner   VARCHAR(254),
 						type    VARCHAR(10) NOT NULL,
 						key     CHAR(36) UNIQUE NOT NULL,
@@ -43,15 +43,15 @@ func migrateDB(db *sql.DB) error {
 						PRIMARY KEY (id, owner)
 					)`,
 					`CREATE TABLE channels (
-						id    SERIAL,
+						id    BIGSERIAL,
 						owner VARCHAR(254),
 						name  TEXT,
 						PRIMARY KEY (id, owner)
 					)`,
 					`CREATE TABLE connections (
-						channel_id    INTEGER,
+						channel_id    BIGINT,
 						channel_owner VARCHAR(254),
-						thing_id     INTEGER,
+						thing_id     BIGINT,
 						thing_owner  VARCHAR(254),
 						FOREIGN KEY (channel_id, channel_owner) REFERENCES channels (id, owner) ON DELETE CASCADE ON UPDATE CASCADE,
 						FOREIGN KEY (thing_id, thing_owner) REFERENCES things (id, owner) ON DELETE CASCADE ON UPDATE CASCADE,
